# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoXidized API
  version: 0.1.0
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
  /:
    get:
      security: []
      responses:
        "302":
          description: Redirects to the embedded Web UI at /ui/.
  /ui/:
    get:
      security: []
      responses:
        "200":
          description: Embedded Web UI shell. API data remains RBAC-gated.
  /healthz:
    get:
      security: []
      responses:
        "200":
          description: Process is alive.
  /readyz:
    get:
      security: []
      responses:
        "200":
          description: Dependencies are ready.
  /metrics:
    get:
      security: []
      responses:
        "200":
          description: Prometheus metrics.
  /auth/oidc/login:
    get:
      security: []
      responses:
        "302":
          description: Redirects to the configured OIDC provider.
        "404":
          description: OIDC login is not enabled.
  /auth/oidc/callback:
    get:
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        "302":
          description: OIDC session cookie created and browser redirected to /ui/.
        "400":
          description: Missing or mismatched OIDC state.
        "403":
          description: OIDC user is not provisioned or lacks roles.
  /auth/logout:
    post:
      responses:
        "200":
          description: Current token revoked and session cookie cleared.
        "401":
          description: Missing or invalid token.
  /api/v1/auth/me:
    get:
      responses:
        "200":
          description: Current principal.
        "401":
          description: Missing or invalid token.
  /api/v1/devices:
    get:
      description: Requires `devices:read`.
      responses:
        "200":
          description: List devices.
        "403":
          description: Permission denied.
  /api/v1/devices/{id}:
    get:
      description: Requires `devices:read`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Device detail.
        "403":
          description: Permission denied.
  /api/v1/devices/{id}/backup:
    post:
      description: Requires `backups:run`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: Backup queued.
        "403":
          description: Permission denied.
  /api/v1/groups/{group}/backup:
    post:
      description: Requires `backups:run`.
      parameters:
        - name: group
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: Group backup queued.
        "403":
          description: Permission denied.
  /api/v1/devices/{id}/configs:
    get:
      description: Requires `configs:read`. Audit metadata excludes config content.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Revision history.
        "403":
          description: Permission denied.
  /api/v1/devices/{id}/configs/latest:
    get:
      description: Requires `configs:read`. Audit metadata excludes config content.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Latest sanitized config.
        "403":
          description: Permission denied.
  /api/v1/devices/{id}/configs/diff:
    get:
      description: Requires `configs:diff`. Audit metadata excludes diff text.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: from
          in: query
          required: true
          schema: { type: string }
        - name: to
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Unified diff.
        "403":
          description: Permission denied.
  /api/v1/jobs:
    get:
      description: Requires `jobs:read`.
      responses:
        "200":
          description: Job list.
        "403":
          description: Permission denied.
  /api/v1/jobs/{id}:
    get:
      description: Requires `jobs:read`.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job detail.
        "403":
          description: Permission denied.
  /api/v1/inventory/reload:
    post:
      description: Requires `inventory:reload`.
      responses:
        "200":
          description: Inventory reloaded.
        "403":
          description: Permission denied.
  /api/v1/drivers:
    get:
      description: Requires `drivers:read`.
      responses:
        "200":
          description: Registered drivers.
        "403":
          description: Permission denied.
  /api/v1/drivers/{name}/test:
    post:
      description: Requires `drivers:test`.
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: Driver test endpoint registered.
        "403":
          description: Permission denied.
  /api/v1/audit/events:
    get:
      description: Requires `audit:read`. Reading audit events is itself audited.
      responses:
        "200":
          description: Audit event list.
        "403":
          description: Permission denied.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: goxidized_session